fix the bug of contract_vol

+++ okex/OKExFutureRest.go
@@ -66,9 +66,9 @@ type okexFutureContract struct {
 	BaseCurrency        string  `json:"base_currency"`          // 交易货币，如：BTC-USD中的BTC ,BTC-USDT中的BTC
 	QuoteCurrency       string  `json:"quote_currency"`         // 计价货币币种，如：BTC-USD中的USD ,BTC-USDT中的USDT
 	SettlementCurrency  string  `json:"settlement_currency"`    // 盈亏结算和保证金币种，如：BTC
-	TickSize            float64 `json:"tick_size,string"`              //下单价格精度
+	TickSize            float64 `json:"tick_size,string"`       //下单价格精度
 	TradeIncrement      float64 `json:"trade_increment,string"` //数量精度
-	ContractVal         int64   `json:"contract_val,string"`    //the contract vol in usd
+	ContractVal         float64 `json:"contract_val,string"`    //the contract vol in usd
 	ContractValCurrency string  `json:"contract_val_currency"`  //合约面值计价币种 如 usd，btc，ltc，etc xrp eos
 	Listing             string  `json:"listing"`                // 上线日期
 	Delivery            string  `json:"delivery"`               // delivery date 交割日期
@@ -86,7 +86,6 @@ func (future *Future) updateFutureContracts() ([]byte, error) {
 	if err != nil {
 		return nil, err
 	}
-
 	SyncTime := time.Now().In(future.config.Location)
 	asia, _ := time.LoadLocation("Asia/Shanghai")
 	futureContracts := FutureContracts{
@@ -147,7 +146,7 @@ func (future *Future) updateFutureContracts() ([]byte, error) {
 			DueTimestamp: dueTime.UnixNano() / int64(time.Millisecond),
 			DueDate:      dueTime.Format(GO_BIRTHDAY),
 
-			UnitAmount:      item.ContractVal,
+			UnitAmount:      int64(item.ContractVal),
 			PricePrecision:  pricePrecision,
 			AmountPrecision: amountPrecision,
 		}
